src/requirements.txt
bleach==3.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
cycler==0.11.0
docutils==0.16
fonttools==4.33.3
idna==2.10
importlib-metadata==3.7.2
keyring==23.0.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
opencv-python==4.6.0.66
packaging==20.9
Pillow==9.1.1
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.2
pywin32-ctypes==0.2.0
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.4.0
six==1.15.0
tqdm==4.59.0
ttkbootstrap==1.9.0
twine==3.3.0
urllib3==1.26.3
webencodings==0.5.1
wincertstore==0.2
zipp==3.4.1
